Refactor of placement of WorkspaceFileOpException, and cleanup in codeRunner.ts
