+++ workspace_server/src/main/java/gov/nasa/jpl/aerie/workspace/server/WorkspaceFileOpException.java
@@ -1,4 +1,4 @@
-package gov.nasa.jpl.aerie.workspace.server.postgres;
+package gov.nasa.jpl.aerie.workspace.server;
 
 public class WorkspaceFileOpException extends Exception {
   public WorkspaceFileOpException(String msg) {
